backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib
bcrypt==3.2.0
python-multipart
pandas
httpx
python-dotenv
uuid
python-slugify
